requirements.txt
apache-airflow>=2.4.0
requests>=2.25.0